requirements.txt
langchain
langchain_community
streamlit
pypdf
langchain_pinecone
pinecone